changed numSacksPerWife primitive to int

+++ Lab2.java
@@ -18,7 +18,7 @@ public class Lab2 {
 	
 	boolean manGoingToStIves = !true;
 
-	short numWives = 7;
+	int numWives = 7;
 
 	int numSacksPerWife = 7;
